Inference Images should be different than training
The Learning Path points people to the same set of images for both Training the Custom Vision Service and using the app afterwards in production. This is incorrect, a different set of images must be used. Also the folder structure of the training set would overwrite half the images if used for inference because they have the same name. It also misses the 0.jpg described in the path. This is a set of creative commons use images to be referred to in the Unit 4. I will update the docs shortly to refer to it.